work through all failing and erroring unit tests and fix as many as possible where the failures are a direct result of the model changes.  This omits fixing errors in subsystems which will be dealt with in later phases of the work

+++ portality/scripts/history_dirs_reports.py
@@ -1,6 +1,6 @@
 import os, json, csv
 
-WORKING_DIR = "/home/richard/tmp/doaj/history/archive/"
+WORKING_DIR = "/home/richard/tmp/doaj/history.20200306/history"
 
 BASE_DIRS = [
     "article_2016_web-pre2019",
@@ -11,6 +11,8 @@ BASE_DIRS = [
     "article_2019_background-pre2019",
     "article_2019_web-2019",
     "article_2019_web-pre2019",
+    "article_2020_background-2019",
+    "article_2020_web-2019",
     "journal_2016_background-pre2019",
     "journal_2017_background-pre2019",
     "journal_2017_web-pre2019",
@@ -18,14 +20,20 @@ BASE_DIRS = [
     "journal_2018_web-pre2019",
     "journal_2019_background-2019",
     "journal_2019_background-pre2019",
-    "journal_2019_web-pre2019"
+    "journal_2019_web-pre2019",
+    "journal_2019_web-2019",
+    "journal_2020_background-2019",
+    "journal_2020_web-2019"
 ]
 
-OUT_DIR = "/home/richard/tmp/doaj/history/archive/"
+OUT_DIR = "/home/richard/tmp/doaj/history.20200306"
 
 
 def history_dirs_reports(working_dir, base_dirs, out_dir):
     for bd in base_dirs:
+        if not os.path.exists(os.path.join(working_dir, bd)):
+            continue
+
         out = os.path.join(out_dir, bd + ".csv")
 
         with open(out, "w", encoding="utf-8") as o:
